workload request 쿼리등 수정

+++ HybridDeployOptimizer/src/main/java/com/kware/policy/task/collector/service/vo/PromMetricPod.java
@@ -78,7 +78,7 @@ public class PromMetricPod extends PromMetricDefault{
 	
 	private String priorityClass; //해당 파드의 priority 정보를 가져와서 뭘할까????? kube_pod_info에 보이는데 일단 이건 나중에
 
-	private Map<String, Integer>  mUsgeGpuMap   = new HashMap<String, Integer>(); //개별 GPU사용량:<GPU 번호, 사용량은 %>
+	private Map<String, Double>  mUsgeGpuMap   = new HashMap<String, Double>(); //개별 GPU사용량:<GPU 번호, 사용량은 %>
 	private Map<String, Long>     mLimitsList   = new HashMap<String, Long>();   //리소스이름: <(cpu, memory, gpu, disk), 값>
 	private Map<String, Long>     mRequestsList = new HashMap<String, Long>(); //리소스이름: <(cpu, memory, gpu, disk), 값>
 
@@ -115,7 +115,7 @@ public class PromMetricPod extends PromMetricDefault{
         }
         
 		try {
-			mUsgeGpuMap.put(result.toString(), Integer.parseInt(vals[len - 1]));
+			mUsgeGpuMap.put(result.toString(), Double.parseDouble(vals[len - 1]));
 		}catch(Exception e) {
 			log.error("setUsageGpu error:{}", _val, e);
 		}
